refactor(cli): split tsserver code-fix fallback builders into fix-family modules (#9413)
Behavior-preserving split of handlers_code_fixes.rs (1899L) to bring it
under the 2000-line cap. Pure code movement; no logic changes.

Split:
- handlers_code_fixes_combined.rs (345L): handle_get_combined_code_fix
- handlers_code_fixes_isolated_decl.rs (229L): TS9010 isolated-decl fix
  (infer_type_for_isolated_decl_initializer + apply_isolated_decl_type_annotation_fix
  + FIX_MISSING_TYPE_ANNOTATION_FIX_ID const)
- handlers_code_fixes_synthetic.rs (+271L, 929L total): six synthetic
  diagnostic helpers (synthetic_implements_interface_diagnostics,
  synthetic_missing_async_suggestion_diagnostic,
  synthetic_add_parameter_names_suggestion_diagnostic,
  synthetic_missing_attributes_suggestion_diagnostic,
  synthetic_add_missing_new_diagnostic,
  synthetic_add_missing_const_diagnostic)
- handlers_code_fixes_imports.rs (+46L, 1550L total): has_potential_auto_import_symbol
- handlers_code_fixes.rs reduced from 1899L to 1046L
