+++ crates/tsz-cli/src/bin/tsz_server/handlers_code_fixes_combined.rs
@@ -1,0 +1,345 @@
+//! `getCombinedCodeFix` handler for tsz-server.
+//!
+//! Contains `handle_get_combined_code_fix` and its supporting logic for
+//! applying a single fix across all diagnostics in a file.
+
+use super::{Server, TsServerRequest, TsServerResponse};
+use tsz::lsp::code_actions::{
+    CodeActionContext, CodeActionKind, CodeActionProvider, CodeFixRegistry,
+};
+use tsz::lsp::position::LineMap;
+
+impl Server {
+    pub(crate) fn handle_get_combined_code_fix(
+        &mut self,
+        seq: u64,
+        request: &TsServerRequest,
+    ) -> TsServerResponse {
+        let file = request
+            .arguments
+            .get("scope")
+            .and_then(|scope| scope.get("args"))
+            .and_then(|args| args.get("file"))
+            .and_then(|v| v.as_str());
+
+        let fix_id = request.arguments.get("fixId").and_then(|v| v.as_str());
+
+        if let (Some(file_path), Some(fix_id)) = (file, fix_id)
+            && let Some((arena, binder, root, content)) = self.parse_and_bind_file(file_path)
+        {
+            let organize_imports_ignore_case = request
+                .arguments
+                .get("preferences")
+                .and_then(|p| p.get("organizeImportsIgnoreCase"))
+                .and_then(serde_json::Value::as_bool)
+                .or_else(|| {
+                    request
+                        .arguments
+                        .get("organizeImportsIgnoreCase")
+                        .and_then(serde_json::Value::as_bool)
+                })
+                .unwrap_or(self.organize_imports_ignore_case);
+            let line_map = LineMap::build(&content);
+            let provider = CodeActionProvider::new(
+                &arena,
+                &binder,
+                &line_map,
+                file_path.to_string(),
+                &content,
+            )
+            .with_organize_imports_ignore_case(organize_imports_ignore_case)
+            .with_new_line_override(self.new_line_character.clone());
+
+            let mut diagnostics = self.get_semantic_diagnostics_full(file_path, &content);
+            diagnostics.extend(
+                self.synthetic_missing_name_expression_diagnostics(file_path, &content, &binder),
+            );
+            let mut seen_diags = rustc_hash::FxHashSet::default();
+            diagnostics
+                .retain(|d| seen_diags.insert((d.code, d.start, d.length, d.message_text.clone())));
+
+            let to_lsp_diag =
+                |d: &tsz::checker::diagnostics::Diagnostic| tsz::lsp::diagnostics::LspDiagnostic {
+                    range: tsz::lsp::position::Range::new(
+                        line_map.offset_to_position(d.start, &content),
+                        line_map.offset_to_position(d.start + d.length, &content),
+                    ),
+                    message: d.message_text.clone(),
+                    code: Some(d.code),
+                    severity: Some(tsz::lsp::diagnostics::DiagnosticSeverity::Error),
+                    source: Some("tsz".to_string()),
+                    related_information: None,
+                    reports_unnecessary: tsz::lsp::diagnostics::is_unnecessary_code(d.code)
+                        .then_some(true),
+                    reports_deprecated: tsz::lsp::diagnostics::is_deprecated_code(d.code)
+                        .then_some(true),
+                };
+            let mut filtered_diagnostics: Vec<tsz::lsp::diagnostics::LspDiagnostic> = diagnostics
+                .iter()
+                .filter(|d| {
+                    CodeFixRegistry::fixes_for_error_code(d.code)
+                        .iter()
+                        .any(|(_, id, _, _)| *id == fix_id)
+                })
+                .map(to_lsp_diag)
+                .collect();
+            if filtered_diagnostics.is_empty() {
+                filtered_diagnostics = diagnostics
+                    .iter()
+                    .filter(|d| {
+                        CodeFixRegistry::fixes_for_error_code(d.code)
+                            .iter()
+                            .any(|(_, id, _, _)| *id == fix_id)
+                    })
+                    .map(to_lsp_diag)
+                    .collect();
+            }
+
+            let auto_import_file_exclude_patterns =
+                Self::extract_auto_import_file_exclude_patterns(request)
+                    .unwrap_or_else(|| self.auto_import_file_exclude_patterns.clone());
+            let auto_import_specifier_exclude_regexes =
+                Self::extract_auto_import_specifier_exclude_regexes(request)
+                    .unwrap_or_else(|| self.auto_import_specifier_exclude_regexes.clone());
+            let import_module_specifier_ending = request
+                .arguments
+                .get("preferences")
+                .and_then(|p| p.get("importModuleSpecifierEnding"))
+                .and_then(|v| v.as_str())
+                .or_else(|| {
+                    request
+                        .arguments
+                        .get("importModuleSpecifierEnding")
+                        .and_then(serde_json::Value::as_str)
+                })
+                .or(self.completion_import_module_specifier_ending.as_deref());
+            let import_module_specifier_preference = request
+                .arguments
+                .get("preferences")
+                .and_then(|p| p.get("importModuleSpecifierPreference"))
+                .and_then(|v| v.as_str())
+                .or_else(|| {
+                    request
+                        .arguments
+                        .get("importModuleSpecifierPreference")
+                        .and_then(serde_json::Value::as_str)
+                })
+                .or(self.import_module_specifier_preference.as_deref());
+            let import_candidates = if fix_id == "fixMissingImport" {
+                self.collect_import_candidates(
+                    file_path,
+                    &filtered_diagnostics,
+                    &auto_import_file_exclude_patterns,
+                    &auto_import_specifier_exclude_regexes,
+                    import_module_specifier_ending,
+                    import_module_specifier_preference,
+                )
+            } else {
+                Vec::new()
+            };
+
+            if fix_id == "fixMissingImport"
+                && let Some(updated_content) =
+                    Self::apply_missing_imports_fix_all(file_path, &content, &import_candidates)
+                && let Some((start_off, end_off, replacement)) =
+                    Self::compute_minimal_edit(&content, &updated_content)
+            {
+                let start_pos = line_map.offset_to_position(start_off, &content);
+                let end_pos = line_map.offset_to_position(end_off, &content);
+                return TsServerResponse {
+                    seq,
+                    msg_type: "response".to_string(),
+                    command: "getCombinedCodeFix".to_string(),
+                    request_seq: request.seq,
+                    success: true,
+                    message: None,
+                    body: Some(serde_json::json!({
+                        "changes": [{
+                            "fileName": file_path,
+                            "textChanges": [{
+                                "start": { "line": start_pos.line + 1, "offset": start_pos.character + 1 },
+                                "end": { "line": end_pos.line + 1, "offset": end_pos.character + 1 },
+                                "newText": replacement
+                            }]
+                        }]
+                    })),
+                };
+            }
+
+            let context = CodeActionContext {
+                diagnostics: filtered_diagnostics,
+                only: Some(vec![CodeActionKind::QuickFix]),
+                import_candidates,
+            };
+
+            let range = tsz::lsp::position::Range::new(
+                tsz::lsp::position::Position::new(0, 0),
+                line_map.offset_to_position(content.len() as u32, &content),
+            );
+
+            let actions = provider.provide_code_actions(root, range, context);
+
+            let mut file_changes_map: rustc_hash::FxHashMap<
+                String,
+                Vec<tsz::lsp::rename::TextEdit>,
+            > = rustc_hash::FxHashMap::default();
+
+            for action in actions {
+                if let Some(edit) = action.edit {
+                    for (fname, edits) in edit.changes {
+                        file_changes_map.entry(fname).or_default().extend(edits);
+                    }
+                }
+            }
+
+            let mut all_changes: Vec<serde_json::Value> = Vec::new();
+            for (fname, edits) in file_changes_map {
+                let mut text_changes = Vec::new();
+                for edit in edits {
+                    text_changes.push(serde_json::json!({
+                        "start": {
+                            "line": edit.range.start.line + 1,
+                            "offset": edit.range.start.character + 1
+                        },
+                        "end": {
+                            "line": edit.range.end.line + 1,
+                            "offset": edit.range.end.character + 1
+                        },
+                        "newText": edit.new_text
+                    }));
+                }
+
+                all_changes.push(serde_json::json!({
+                    "fileName": fname,
+                    "textChanges": text_changes
+                }));
+            }
+
+            if fix_id == "annotateWithTypeFromJSDoc"
+                && let Some(updated_content) =
+                    Self::apply_simple_jsdoc_annotation_fallback(&content)
+                && let Some((start_off, end_off, replacement)) =
+                    Self::compute_minimal_edit(&content, &updated_content)
+            {
+                let start_pos = line_map.offset_to_position(start_off, &content);
+                let end_pos = line_map.offset_to_position(end_off, &content);
+                all_changes.clear();
+                all_changes.push(serde_json::json!({
+                    "fileName": file_path,
+                    "textChanges": [{
+                        "start": { "line": start_pos.line + 1, "offset": start_pos.character + 1 },
+                        "end": { "line": end_pos.line + 1, "offset": end_pos.character + 1 },
+                        "newText": replacement
+                    }]
+                }));
+            }
+
+            if all_changes.is_empty()
+                && fix_id == "addConvertToUnknownForNonOverlappingTypes"
+                && let Some(updated_content) = Self::apply_unknown_conversion_fallback(&content)
+            {
+                let end_pos = line_map.offset_to_position(content.len() as u32, &content);
+                all_changes.push(serde_json::json!({
+                    "fileName": file_path,
+                    "textChanges": [{
+                        "start": { "line": 1, "offset": 1 },
+                        "end": { "line": end_pos.line + 1, "offset": end_pos.character + 1 },
+                        "newText": updated_content
+                    }]
+                }));
+            }
+
+            if all_changes.is_empty()
+                && fix_id == "addNameToNamelessParameter"
+                && let Some(updated_content) =
+                    Self::apply_add_names_to_nameless_parameters_fallback(&content)
+            {
+                let end_pos = line_map.offset_to_position(content.len() as u32, &content);
+                all_changes.push(serde_json::json!({
+                    "fileName": file_path,
+                    "textChanges": [{
+                        "start": { "line": 1, "offset": 1 },
+                        "end": { "line": end_pos.line + 1, "offset": end_pos.character + 1 },
+                        "newText": updated_content
+                    }]
+                }));
+            }
+
+            if all_changes.is_empty()
+                && fix_id == "fixMissingAttributes"
+                && let Some(updated_content) = Self::apply_missing_attributes_fallback(&content)
+            {
+                let end_pos = line_map.offset_to_position(content.len() as u32, &content);
+                all_changes.push(serde_json::json!({
+                    "fileName": file_path,
+                    "textChanges": [{
+                        "start": { "line": 1, "offset": 1 },
+                        "end": { "line": end_pos.line + 1, "offset": end_pos.character + 1 },
+                        "newText": updated_content
+                    }]
+                }));
+            }
+
+            if all_changes.is_empty()
+                && fix_id == "addMissingAwait"
+                && let Some((_, updated_content)) =
+                    Self::apply_add_missing_await_fallback(&content, true)
+            {
+                let end_pos = line_map.offset_to_position(content.len() as u32, &content);
+                all_changes.push(serde_json::json!({
+                    "fileName": file_path,
+                    "textChanges": [{
+                        "start": { "line": 1, "offset": 1 },
+                        "end": { "line": end_pos.line + 1, "offset": end_pos.character + 1 },
+                        "newText": updated_content
+                    }]
+                }));
+            }
+
+            if fix_id == "addMissingConst"
+                && let Some(updated_content) =
+                    Self::apply_add_missing_const_fix_all_fallback(&content)
+            {
+                let end_pos = line_map.offset_to_position(content.len() as u32, &content);
+                all_changes.clear();
+                all_changes.push(serde_json::json!({
+                    "fileName": file_path,
+                    "textChanges": [{
+                        "start": { "line": 1, "offset": 1 },
+                        "end": { "line": end_pos.line + 1, "offset": end_pos.character + 1 },
+                        "newText": updated_content
+                    }]
+                }));
+            }
+
+            if all_changes.is_empty()
+                && fix_id == "addMissingNewOperator"
+                && let Some(updated_content) = Self::apply_add_missing_new_all_fallback(&content)
+            {
+                let end_pos = line_map.offset_to_position(content.len() as u32, &content);
+                all_changes.push(serde_json::json!({
+                    "fileName": file_path,
+                    "textChanges": [{
+                        "start": { "line": 1, "offset": 1 },
+                        "end": { "line": end_pos.line + 1, "offset": end_pos.character + 1 },
+                        "newText": updated_content
+                    }]
+                }));
+            }
+
+            return TsServerResponse {
+                seq,
+                msg_type: "response".to_string(),
+                command: "getCombinedCodeFix".to_string(),
+                request_seq: request.seq,
+                success: true,
+                message: None,
+                body: Some(serde_json::json!({
+                    "changes": all_changes
+                })),
+            };
+        }
+
+        self.success_response(seq, request, Some(serde_json::json!({"changes": []})))
+    }
+}
